fix: separate game page metadata from client component
